+++ Userland/Libraries/LibJS/Runtime/ArrayIteratorPrototype.cpp
@@ -9,7 +9,7 @@
 #include <LibJS/Runtime/ArrayIteratorPrototype.h>
 #include <LibJS/Runtime/Error.h>
 #include <LibJS/Runtime/GlobalObject.h>
-#include <LibJS/Runtime/IteratorOperations.h>
+#include <LibJS/Runtime/Iterator.h>
 #include <LibJS/Runtime/TypedArray.h>
 
 namespace JS {
