+++ Userland/Libraries/LibJS/Runtime/Intl/ListFormat.cpp
@@ -7,7 +7,7 @@
 #include <LibJS/Runtime/Array.h>
 #include <LibJS/Runtime/GlobalObject.h>
 #include <LibJS/Runtime/Intl/ListFormat.h>
-#include <LibJS/Runtime/IteratorOperations.h>
+#include <LibJS/Runtime/Iterator.h>
 #include <LibJS/Runtime/ThrowableStringBuilder.h>
 
 namespace JS::Intl {
