+++ Userland/Libraries/LibJS/Runtime/MapIteratorPrototype.cpp
@@ -8,7 +8,7 @@
 #include <LibJS/Runtime/Array.h>
 #include <LibJS/Runtime/Error.h>
 #include <LibJS/Runtime/GlobalObject.h>
-#include <LibJS/Runtime/IteratorOperations.h>
+#include <LibJS/Runtime/Iterator.h>
 #include <LibJS/Runtime/MapIteratorPrototype.h>
 
 namespace JS {
