+++ Userland/Libraries/LibJS/Runtime/RegExpStringIteratorPrototype.cpp
@@ -5,7 +5,7 @@
  */
 
 #include <LibJS/Runtime/GlobalObject.h>
-#include <LibJS/Runtime/IteratorOperations.h>
+#include <LibJS/Runtime/Iterator.h>
 #include <LibJS/Runtime/RegExpPrototype.h>
 #include <LibJS/Runtime/RegExpStringIteratorPrototype.h>
 #include <LibJS/Runtime/Utf16String.h>
