+++ Userland/Libraries/LibJS/Runtime/SetConstructor.cpp
@@ -7,7 +7,7 @@
 #include <LibJS/Runtime/AbstractOperations.h>
 #include <LibJS/Runtime/Error.h>
 #include <LibJS/Runtime/GlobalObject.h>
-#include <LibJS/Runtime/IteratorOperations.h>
+#include <LibJS/Runtime/Iterator.h>
 #include <LibJS/Runtime/Set.h>
 #include <LibJS/Runtime/SetConstructor.h>
