+++ Userland/Libraries/LibJS/Runtime/SetIteratorPrototype.cpp
@@ -8,7 +8,7 @@
 #include <LibJS/Runtime/Array.h>
 #include <LibJS/Runtime/Error.h>
 #include <LibJS/Runtime/GlobalObject.h>
-#include <LibJS/Runtime/IteratorOperations.h>
+#include <LibJS/Runtime/Iterator.h>
 #include <LibJS/Runtime/SetIteratorPrototype.h>
 
 namespace JS {
