+++ Userland/Libraries/LibJS/Runtime/StringIteratorPrototype.cpp
@@ -7,7 +7,7 @@
 #include <AK/TypeCasts.h>
 #include <LibJS/Runtime/Error.h>
 #include <LibJS/Runtime/GlobalObject.h>
-#include <LibJS/Runtime/IteratorOperations.h>
+#include <LibJS/Runtime/Iterator.h>
 #include <LibJS/Runtime/StringIteratorPrototype.h>
 #include <LibJS/Runtime/ThrowableStringBuilder.h>
