+++ Userland/Libraries/LibJS/Runtime/TypedArray.cpp
@@ -12,7 +12,7 @@
 #include <LibJS/Runtime/ArrayBufferConstructor.h>
 #include <LibJS/Runtime/Completion.h>
 #include <LibJS/Runtime/GlobalObject.h>
-#include <LibJS/Runtime/IteratorOperations.h>
+#include <LibJS/Runtime/Iterator.h>
 #include <LibJS/Runtime/TypedArray.h>
 #include <LibJS/Runtime/TypedArrayConstructor.h>
