+++ Userland/Libraries/LibJS/Runtime/WeakSetConstructor.cpp
@@ -7,7 +7,7 @@
 #include <LibJS/Runtime/AbstractOperations.h>
 #include <LibJS/Runtime/Error.h>
 #include <LibJS/Runtime/GlobalObject.h>
-#include <LibJS/Runtime/IteratorOperations.h>
+#include <LibJS/Runtime/Iterator.h>
 #include <LibJS/Runtime/WeakSet.h>
 #include <LibJS/Runtime/WeakSetConstructor.h>
