+++ Userland/Libraries/LibWeb/HTML/CustomElements/CustomElementRegistry.cpp
@@ -5,7 +5,7 @@
  */
 
 #include <LibJS/Runtime/FunctionObject.h>
-#include <LibJS/Runtime/IteratorOperations.h>
+#include <LibJS/Runtime/Iterator.h>
 #include <LibWeb/Bindings/CustomElementRegistryPrototype.h>
 #include <LibWeb/DOM/Document.h>
 #include <LibWeb/DOM/ElementFactory.h>
